refactor(assets): isolate platform and local file logic for CLI tools
Extracts platform identifier generation into a dedicated CliPlatform
utility, ensuring consistent mapping to manifest keys (e.g., linux-x64).
Also isolates all local file resolution, version parsing, and cache
cleanup logic from DefaultAppLandDownloadService into a new
LocalAssetRepository.

This purely structural change prepares the download service to
transition from GitHub release-based discovery to a manifest-driven
strategy.
