+++ plugin-core/src/main/java/appland/cli/DefaultAppLandDownloadService.java
@@ -6,14 +6,11 @@ import appland.settings.DownloadSettings;
 import appland.utils.GsonUtils;
 import com.google.gson.JsonArray;
 import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.application.PathManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.progress.Task;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.util.SystemInfo;
-import com.intellij.openapi.util.Version;
 import com.intellij.openapi.util.io.NioFiles;
 import com.intellij.util.Urls;
 import com.intellij.util.io.HttpRequests;
@@ -23,13 +20,9 @@ import org.jetbrains.annotations.Nullable;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.util.*;
-import java.util.stream.Collectors;
-
-import static appland.cli.CliTools.currentArch;
-import static appland.cli.CliTools.currentPlatform;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Downloads are stored as "appland-downloads/$NAME/$VERSION/$NAME-$OS-$ARCH", e.g. "appland-downloads/appmap/1.2.3/appmap-linux-x64".
@@ -42,13 +35,13 @@ public class DefaultAppLandDownloadService implements AppLandDownloadService {
 
     @Override
     public @Nullable Path getDownloadFilePath(@NotNull CliTool type, @NotNull String platform, @NotNull String arch) {
-        var version = findLatestDownloadedVersion(type);
+        var version = LocalAssetRepository.findLatestDownloadedVersion(type);
         if (version == null) {
             return null;
         }
 
         var unitTestMode = ApplicationManager.getApplication().isUnitTestMode();
-        return getExecutableFilePath(type, version, platform, arch, unitTestMode);
+        return LocalAssetRepository.getExecutableFilePath(type, version, platform, arch, unitTestMode);
     }
 
     @Override
@@ -63,11 +56,11 @@ public class DefaultAppLandDownloadService implements AppLandDownloadService {
         var unitTestMode = ApplicationManager.getApplication().isUnitTestMode();
 
         try {
-            var platform = currentPlatform();
-            var arch = currentArch();
+            var platform = CliPlatform.currentPlatform();
+            var arch = CliPlatform.currentArch();
 
             // the path to the downloaded executable
-            var targetFilePath = getExecutableFilePath(type, version, platform, arch, unitTestMode);
+            var targetFilePath = LocalAssetRepository.getExecutableFilePath(type, version, platform, arch, unitTestMode);
             // the path where the in-progress download is stored
             var downloadTargetFilePath = targetFilePath.resolveSibling(targetFilePath.getFileName().toString() + ".download");
 
@@ -89,7 +82,7 @@ public class DefaultAppLandDownloadService implements AppLandDownloadService {
             Files.move(downloadTargetFilePath, targetFilePath, StandardCopyOption.ATOMIC_MOVE);
 
             // after a successful download, remove previous downloads, make sure to keep the new download
-            removeOtherVersions(type, version, unitTestMode);
+            LocalAssetRepository.removeOtherVersions(type, version, unitTestMode);
 
             notifyDownloadFinished(type, AppMapDownloadStatus.Successful);
             return AppMapDownloadStatus.Successful;
@@ -97,7 +90,7 @@ public class DefaultAppLandDownloadService implements AppLandDownloadService {
             LOG.debug("Error downloading CLI binary", e);
 
             // cleanup failed download, if available
-            var downloadDir = getVersionDownloadDirectory(type, version, unitTestMode);
+            var downloadDir = LocalAssetRepository.getVersionDownloadDirectory(type, version, unitTestMode);
             try {
                 if (Files.isDirectory(downloadDir)) {
                     NioFiles.deleteRecursively(downloadDir);
@@ -135,90 +128,15 @@ public class DefaultAppLandDownloadService implements AppLandDownloadService {
         }
     }
 
+    @Override
     public @Nullable String findLatestDownloadedVersion(@NotNull CliTool type) {
-        var directory = getToolDownloadDirectory(type, ApplicationManager.getApplication().isUnitTestMode());
-        if (!Files.isDirectory(directory)) {
-            return null;
-        }
-
-        var candidates = findVersionDownloadDirectories(directory);
-        if (candidates.isEmpty()) {
-            return null;
-        }
-
-        return candidates
-                .stream()
-                .map(file -> file.getFileName().toString())
-                .max(Comparator.naturalOrder())
-                .orElse(null);
-    }
-
-    private boolean isDownloaded(@NotNull CliTool type, @NotNull String version, boolean unitTestMode) {
-        return isExecutableBinary(getExecutableFilePath(type, version, currentPlatform(), currentArch(), unitTestMode));
-    }
-
-    private boolean isExecutableBinary(@NotNull Path file) {
-        return Files.isRegularFile(file) && (!SystemInfo.isUnix || Files.isExecutable(file));
-    }
-
-    static @NotNull List<Path> findVersionDownloadDirectories(@NotNull Path parentDirectory) {
-        try (var stream = Files.list(parentDirectory)) {
-            return stream
-                    .filter(file -> Files.isDirectory(file) && Version.parseVersion(file.getFileName().toString()) != null)
-                    .collect(Collectors.toList());
-        } catch (Exception e) {
-            LOG.debug("Error listing download directory entries: " + parentDirectory, e);
-            return Collections.emptyList();
-        }
-    }
-
-    static void removeOtherVersions(@NotNull CliTool type, @NotNull String keepVersion, boolean unitTestMode) {
-        removeOtherVersions(getToolDownloadDirectory(type, unitTestMode), keepVersion);
-    }
-
-    @NotNull
-    static List<Path> removeOtherVersions(@NotNull Path directory, @NotNull String keepVersion) {
-        var toRemove = findVersionDownloadDirectories(directory)
-                .stream()
-                .filter(path -> !keepVersion.equals(path.getFileName().toString()))
-                .collect(Collectors.toList());
-
-        if (!toRemove.isEmpty()) {
-            for (var path : toRemove) {
-                try {
-                    NioFiles.deleteRecursively(path);
-                } catch (IOException e) {
-                    LOG.debug("Error deleting download directory: " + path, e);
-                }
-            }
-        }
-        return toRemove;
-    }
-
-    static @NotNull Path getExecutableFilePath(@NotNull CliTool type,
-                                               @NotNull String version,
-                                               @NotNull String platform,
-                                               @NotNull String arch,
-                                               boolean unitTestMode) {
-        return getVersionDownloadDirectory(type, version, unitTestMode).resolve(type.getBinaryName(platform, arch));
-    }
-
-    static @NotNull Path getVersionDownloadDirectory(@NotNull CliTool type, @NotNull String version, boolean unitTestMode) {
-        return getToolDownloadDirectory(type, unitTestMode).resolve(version);
-    }
-
-    @NotNull
-    static Path getToolDownloadDirectory(@NotNull CliTool type, boolean unitTestMode) {
-        var basePath = unitTestMode || "true".equals(System.getProperty("appmap.sandbox"))
-                ? Paths.get(PathManager.getTempPath()).resolve("appland-downloads")
-                : Paths.get(PathManager.getDefaultSystemPathFor("appland-plugin"));
-        return basePath.resolve(type.getId());
+        return LocalAssetRepository.findLatestDownloadedVersion(type);
     }
 
     private void downloadTool(@NotNull Project project, @NotNull CliTool type) {
         LOG.debug("Downloading AppMap CLI tool: " + type);
         var latestVersion = fetchLatestRemoteVersion(type);
-        if (latestVersion != null && !isDownloaded(type, latestVersion, ApplicationManager.getApplication().isUnitTestMode())) {
+        if (latestVersion != null && !LocalAssetRepository.isDownloaded(type, latestVersion, CliPlatform.currentPlatform(), CliPlatform.currentArch(), ApplicationManager.getApplication().isUnitTestMode())) {
             var title = AppMapBundle.get("cliDownload.progress.title", type.getPresentableName());
             new Task.Backgroundable(project, title, true) {
                 @Override
